- 수행 목록 -- 리뷰 조회 로직 구현 (+ 페이징 처리)

+++ src/main/java/com/jhsfully/reservation/service/ReviewService.java
@@ -15,7 +15,7 @@ public interface ReviewService {
 
     void deleteReview(Long memberId, Long reviewId);
 
-    List<ReviewDto.ReviewResponse> getReviewsForUser(Long memberId, Long pageIndex);
+    List<ReviewDto.ReviewResponse> getReviewsForUser(Long memberId, int pageIndex);
 
-    List<ReviewDto.ReviewResponse> getReviewsForShop(Long shopId, Long pageIndex);
+    List<ReviewDto.ReviewResponse> getReviewsForShop(Long shopId, int pageIndex);
 }
